feat: MoreInfosField cmpt
